Update jurisdiction references from DIVORCE to FR in validation rules and tests
