+++ src/test/java/uk/gov/hmcts/reform/orgrolemapping/domain/service/DroolFrAdminOrgRoleMappingTest.java
@@ -5,6 +5,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
+import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.mockito.junit.jupiter.MockitoExtension;
 import uk.gov.hmcts.reform.orgrolemapping.domain.model.CaseWorkerAccessProfile;
@@ -16,20 +17,26 @@ import uk.gov.hmcts.reform.orgrolemapping.domain.model.enums.GrantType;
 import uk.gov.hmcts.reform.orgrolemapping.domain.model.enums.Jurisdiction;
 import uk.gov.hmcts.reform.orgrolemapping.domain.model.enums.RoleCategory;
 import uk.gov.hmcts.reform.orgrolemapping.domain.model.enums.RoleType;
+import uk.gov.hmcts.reform.orgrolemapping.helper.TestDataBuilder;
 import uk.gov.hmcts.reform.orgrolemapping.domain.model.enums.crd.JobTitle;
 import uk.gov.hmcts.reform.orgrolemapping.helper.UserAccessProfileBuilder;
 
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.function.Function;
 import java.util.stream.Collectors;
+import java.util.stream.IntStream;
 import java.util.stream.Stream;
 
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.collection.IsIterableContainingInAnyOrder.containsInAnyOrder;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNull;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 @Slf4j
 @ExtendWith(MockitoExtension.class)
@@ -37,11 +44,23 @@ class DroolFrAdminOrgRoleMappingTest extends DroolBase {
 
     private static final String PRIMARY_LOCATION_ID = UserAccessProfileBuilder.PRIMARY_LOCATION_ID;
     private static final String REGION_ID = "LDN";
-    private static final String JURISDICTION = Jurisdiction.DIVORCE.getName();
+    private static final String JURISDICTION = Jurisdiction.FR.getName();
     private static final String FEATURE_FLAG_PREFIX = "FR";
 
     private static final String WORK_TYPES_HEARING = "routine_work, review_case";
     private static final String WORK_TYPES_ACCESS_REQUESTS = "access_requests";
+        private static final String WORK_TYPES_CTSC_TEAM_LEADER =
+            "hearing_work, routine_work, decision_making_work, applications, review_case, evidence";
+        private static final String WORK_TYPES_CTSC =
+            "hearing_work, routine_work, applications, review_case, evidence";
+
+        private static final Map<String, String> STAFF_ROLE_WORK_TYPES = new HashMap<>();
+
+        static {
+        STAFF_ROLE_WORK_TYPES.put(RoleName.CTSC_TEAM_LEADER, WORK_TYPES_CTSC_TEAM_LEADER);
+        STAFF_ROLE_WORK_TYPES.put(RoleName.CTSC, WORK_TYPES_CTSC);
+        STAFF_ROLE_WORK_TYPES.put(RoleName.SPECIFIC_ACCESS_APPROVER_CTSC, WORK_TYPES_ACCESS_REQUESTS);
+        }
 
     private record ExpectedRole(String roleName,
                                 RoleCategory roleCategory,
@@ -63,7 +82,7 @@ class DroolFrAdminOrgRoleMappingTest extends DroolBase {
     private static ExpectedRole standardAdminRole(String roleName, String workTypes) {
         return new ExpectedRole(roleName, RoleCategory.ADMIN,
                 Classification.PUBLIC, GrantType.STANDARD, false,
-                JURISDICTION, PRIMARY_LOCATION_ID, null, workTypes);
+                JURISDICTION, PRIMARY_LOCATION_ID, REGION_ID, workTypes);
     }
 
     private static ExpectedRole standardAdminRoleWithRegion(String roleName, String workTypes) {
@@ -75,13 +94,13 @@ class DroolFrAdminOrgRoleMappingTest extends DroolBase {
     private static ExpectedRole taskSupervisor() {
         return new ExpectedRole(RoleName.TASK_SUPERVISOR, RoleCategory.ADMIN,
                 Classification.PUBLIC, GrantType.STANDARD, false,
-                JURISDICTION, PRIMARY_LOCATION_ID, null, null);
+                JURISDICTION, PRIMARY_LOCATION_ID, REGION_ID, null);
     }
 
     private static ExpectedRole caseAllocator() {
         return new ExpectedRole(RoleName.CASE_ALLOCATOR, RoleCategory.ADMIN,
                 Classification.PUBLIC, GrantType.STANDARD, false,
-                JURISDICTION, PRIMARY_LOCATION_ID, null, null);
+                JURISDICTION, PRIMARY_LOCATION_ID, REGION_ID, null);
     }
 
     private static final List<ExpectedRole> EXPECTED_ROLES_HEARING_CENTRE_TEAM_LEADER = List.of(
@@ -144,7 +163,7 @@ class DroolFrAdminOrgRoleMappingTest extends DroolBase {
         allProfiles.clear();
 
         CaseWorkerAccessProfile cap = UserAccessProfileBuilder.buildUserAccessProfileForRoleId5();
-        cap.setServiceCode(Jurisdiction.DIVORCE.getServiceCodes().getFirst());
+        cap.setServiceCode(Jurisdiction.FR.getServiceCodes().getFirst());
         cap.setSuspended(false);
         cap.setRoleId(jobTitle.getRoleId());
         cap.setRoleName(jobTitle.getRoleName());
@@ -229,7 +248,7 @@ class DroolFrAdminOrgRoleMappingTest extends DroolBase {
         allProfiles.clear();
 
         CaseWorkerAccessProfile cap = UserAccessProfileBuilder.buildUserAccessProfileForRoleId5();
-        cap.setServiceCode(Jurisdiction.DIVORCE.getServiceCodes().getFirst());
+        cap.setServiceCode(Jurisdiction.FR.getServiceCodes().getFirst());
         cap.setSuspended(false);
         cap.setRoleId(JobTitle.HEARING_CENTRE_TEAM_LEADER.getRoleId());
         cap.setRoleName(JobTitle.HEARING_CENTRE_TEAM_LEADER.getRoleName());
@@ -242,5 +261,103 @@ class DroolFrAdminOrgRoleMappingTest extends DroolBase {
             buildExecuteKieSession(getAllFeatureFlagsToggleByJurisdiction(FEATURE_FLAG_PREFIX, false));
 
         assertEquals(0, roleAssignments.size());
+    }
+
+    @ParameterizedTest
+    @CsvSource({
+        "9,'ctsc-team-leader,ctsc,hmcts-ctsc,specific-access-approver-ctsc',N,N",
+        "9,'ctsc-team-leader,ctsc,hmcts-ctsc,specific-access-approver-ctsc,task-supervisor',Y,N",
+        "9,'ctsc-team-leader,ctsc,hmcts-ctsc,specific-access-approver-ctsc,case-allocator',N,Y",
+        "9,'ctsc-team-leader,ctsc,hmcts-ctsc,specific-access-approver-ctsc,task-supervisor,case-allocator',Y,Y",
+        "10,'ctsc,hmcts-ctsc',N,N",
+        "10,'ctsc,hmcts-ctsc,task-supervisor',Y,N",
+        "10,'ctsc,hmcts-ctsc,case-allocator',N,Y",
+        "10,'ctsc,hmcts-ctsc,task-supervisor,case-allocator',Y,Y"
+    })
+    void shouldReturnDivorceCtscMappings(String roleId,
+                                         String expectedRoles,
+                                         String taskSupervisorFlag,
+                                         String caseAllocatorFlag) {
+        allProfiles.clear();
+
+        CaseWorkerAccessProfile cap = UserAccessProfileBuilder.buildUserAccessProfileForRoleId2();
+        cap.setServiceCode(Jurisdiction.FR.getServiceCodes().getFirst());
+        cap.setSuspended(false);
+        cap.setRoleId(roleId);
+        cap.setTaskSupervisorFlag(taskSupervisorFlag);
+        cap.setCaseAllocatorFlag(caseAllocatorFlag);
+        cap.setRegionId(REGION_ID);
+        allProfiles.add(cap);
+
+        List<RoleAssignment> roleAssignments =
+            buildExecuteKieSession(getAllFeatureFlagsToggleByJurisdiction(FEATURE_FLAG_PREFIX, true));
+
+        String[] roleNames = expectedRoles.isBlank() ? new String[0] : expectedRoles.split(",");
+        assertEquals(roleNames.length, roleAssignments.size());
+        assertThat(roleAssignments.stream().map(RoleAssignment::getRoleName).collect(Collectors.toList()),
+                containsInAnyOrder(roleNames));
+
+        roleAssignments.forEach(r -> {
+            assertEquals(RoleCategory.CTSC, r.getRoleCategory());
+            assertEquals(RoleType.ORGANISATION, r.getRoleType());
+            assertEquals(cap.getId(), r.getActorId());
+        });
+
+        roleAssignments.forEach(r -> {
+            if (RoleName.HMCTS_CTSC.equals(r.getRoleName())) {
+                assertNull(r.getAttributes().get(Attributes.Name.JURISDICTION));
+                assertEquals(Classification.PRIVATE, r.getClassification());
+                assertEquals(GrantType.BASIC, r.getGrantType());
+            } else {
+                assertEquals(JURISDICTION, r.getAttributes().get(Attributes.Name.JURISDICTION).asText());
+                assertEquals(Classification.PUBLIC, r.getClassification());
+                assertEquals(GrantType.STANDARD, r.getGrantType());
+                assertEquals(cap.getPrimaryLocationId(), r.getAttributes().get(Attributes.Name.PRIMARY_LOCATION).asText());
+            }
+            if (STAFF_ROLE_WORK_TYPES.containsKey(r.getRoleName())) {
+                assertEquals(STAFF_ROLE_WORK_TYPES.get(r.getRoleName()),
+                        r.getAttributes().get(Attributes.Name.WORK_TYPES).asText());
+            } else {
+                assertFalse(r.getAttributes().containsKey(Attributes.Name.WORK_TYPES));
+            }
+        });
+    }
+
+    @Test
+    void shouldNotReturnDivorceRolesWhenFeatureFlagIsOff() {
+        allProfiles.clear();
+
+        CaseWorkerAccessProfile cap = UserAccessProfileBuilder.buildUserAccessProfileForRoleId2();
+        cap.setServiceCode(Jurisdiction.FR.getServiceCodes().getFirst());
+        cap.setSuspended(false);
+        cap.setRoleId(JobTitle.CTSC_TEAM_LEADER.getRoleId());
+        cap.setTaskSupervisorFlag("Y");
+        cap.setCaseAllocatorFlag("Y");
+        cap.setRegionId(REGION_ID);
+        allProfiles.add(cap);
+
+        List<RoleAssignment> roleAssignments =
+            buildExecuteKieSession(getAllFeatureFlagsToggleByJurisdiction(FEATURE_FLAG_PREFIX, false));
+
+        assertTrue(roleAssignments.isEmpty());
+    }
+
+    @Test
+    void shouldNotReturnOrgRolesForCaseWorkerWithSuspendedProfile() {
+        allProfiles.clear();
+        IntStream.range(1, 11).forEach(roleId ->
+                allProfiles.add(TestDataBuilder.buildUserAccessProfile(String.valueOf(roleId), true)));
+
+        allProfiles.forEach(userAccessProfile -> {
+            userAccessProfile.setServiceCode(Jurisdiction.FR.getServiceCodes().getFirst());
+            userAccessProfile.setCaseAllocatorFlag("Y");
+            userAccessProfile.setTaskSupervisorFlag("Y");
+            userAccessProfile.setRegionId(REGION_ID);
+        });
+
+        List<RoleAssignment> roleAssignments =
+            buildExecuteKieSession(getAllFeatureFlagsToggleByJurisdiction(FEATURE_FLAG_PREFIX, true));
+
+        assertTrue(roleAssignments.isEmpty());
     }
 }
